requirements.txt
# pip install -r requirements.txt
# pip freeze > requirements.txt

Django
django-extensions
django-taggit
django-model-utils
Unipath

# beautifulsoup4
# bleach
# Scrapy

# django-imagekit
# Pillow

# django-froala-editor
